+++ module/dice.js
@@ -1,5 +1,9 @@
 import { ChatMessagePF } from "./sidebar/chat-message.js";
 
+export const formulaHasDice = function(formula) {
+  return formula.match(/[0-9\)][dD]/) || formula.match(/[dD][0-9\(]/);
+};
+
 export class DicePF {
 
   /**