+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Added safeguard to avoid an infinite loop in the formula for charges on a charged item
 - Certain fields like currency fields didn't accept a number-only value, and didn't apply by clicking on something else
+- Subtracting hit points from a token which has temporary hit points resulted in a faulty value for hit points
 
 ### Changelog